requirements.txt
dnspython==2.4.2
py4j==0.10.9.7
pymongo==4.4.1
pyspark==3.4.1
python-dotenv==1.0.0
setuptools==68.0.0
spark==0.2.1
urllib3==2.0.4
wheel==0.41.1
